Move checkboxes to new file, performance impr
